+++ packages/styleguide/src/lib/Atoms/Icons/constants.tsx
@@ -26,6 +26,7 @@ export const VENDOR_ICONS = pick(
     'MediumIcon',
     'OpenAiFullIcon',
     'OpenAiIcon',
+    'PercipioIcon',
     'RailsIcon',
     'ReactIcon',
     'RedditIcon',